4/1/22 - Scoreboard update
Another update. Fixed empty input method. Edited position of cin.clear and cin.ignore.
Note: If user types anything after inputting their name, it will appear the next time they are asked to enter a value. If this included a new line, that text is used instead of asking for a name.